Refactor: add ts configuration files
